Hackathon Project
Has two main working functions: 

Iterate: 
LED Light is turned on. Move over in the x or y direction, and if you go over the edge, you go to next row or column.

Traverse: 
LED Light is turned on. Moe around in x or y direction. If you are on an edge you can only move away or along it. 

Control X or Y movement with a joypad, and switch between modes with a fifth button.